Refactoring Protobuf schema parser into separate module.
